requirements.txt
aiohttp
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ecdsa==0.19.1
filelock==3.20.3
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
legendary-gl==0.20.34
MarkupSafe==3.0.3
ms_cv==0.1.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
psnawp==3.0.1
pycountry==24.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==81.0.0
six==1.17.0
tabulate==0.9.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.3
xbox-webapi==2.1.0
